requirements.txt
Django>=5.0
djangorestframework>=3.15
twilio>=9.0.0
pymongo>=4.6.0
python-dotenv>=1.0.1

numpy>=1.26.4
pandas>=2.2.0
scikit-learn>=1.4.0

nltk>=3.8.1
transformers>=4.40.0
torch>=2.2.0
tensorflow>=2.16.0

joblib>=1.3.2
requests>=2.31.0
beautifulsoup4>=4.12.3
newspaper3k>=0.2.8
textblob>=0.18.0
spacy>=3.7.0

gunicorn>=21.2.0
celery>=5.4.0
redis>=5.0.0
